Documentation and example for Go >= 1.17
