+++ examples/go/README.md
@@ -1,38 +1,23 @@
-# Golang Runtime with Dep as Dependencies Manager
+# Go Runtime (>= 1.17)
 
 ## Requirements
 
-Every handler must be in its own package, identified by `package main`, and exporting a main function with the following `lambda.Start` statement:
-```go
-// Must Always be package main
-package main
+- your code must be a valid Go module: a `go.mod` file is expected at the root
+- your handler function should be in a file at the root of your module
+- your handler must be exported: `Handle` is correct, `handle` is not
+- `main` package is reserved: you must not have any package named `main` in your module
 
-import (
-	"encoding/json"
-	"http"
-	
-  // Import both of these packages
-	"github.com/scaleway/scaleway-functions-go/events"
-	"github.com/scaleway/scaleway-functions-go/lambda"
-)
+Example of a right structure:
 
-// Handler - Your handler function, uses APIGatewayProxy event type as your function will always get HTTP formatted events, even for CRON
-func Handler(req events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
-	return events.APIGatewayProxyResponse{
-		Body:       "Your response",
-		StatusCode: http.StatusOK,
-	}, nil
-}
-
-// Main function is mandatory -> Must call lambda.Start(yourHandler) otherwhise your handler will not be called properly.
-func main() {
-	lambda.Start(Handler)
-}
+```
+.
+├── go.mod        # your go.mod defines your module
+├── go.sum        # not always necessary 
+├── myfunc.go     # your handler method (exported) must be defined here
+└── subpackage    # you can have subpackages
+    └── hello.go  # with files inside
 ```
 
 ## Run
 
-Additonnaly you may run `go mod vendor`.
-
-We are building `golang` binaries inside our APIs, although we need you to package all your code with `vendors` (dependencies), so a `go build` command would work properly without prior installations on our side.
-
+If your code depends on private dependencies, you will need to run `go mod vendor` before deploying your function.
