+++ examples/go/go.mod
@@ -1,5 +1,3 @@
 module myhandler
 
 go 1.18
-
-require github.com/scaleway/scaleway-functions-go v0.0.0-20211103103919-280251df088b
