+++ examples/go113/go.mod
@@ -1,0 +1,5 @@
+module myhandler
+
+go 1.13
+
+require github.com/scaleway/scaleway-functions-go v0.0.0-20211103103919-280251df088b
