+++ examples/go113/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "golang-scaleway-starter",
+  "version": "1.0.0",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {},
+  "devDependencies": {
+    "serverless-scaleway-functions": "^0.3.1"
+  },
+  "description": ""
+}
